Seperate the onepage address, payment, order-summary, cart-summary, shipping component and code refactor.
