+++ packages/Webkul/Checkout/src/Cart.php
@@ -866,14 +866,14 @@ class Cart
     ): void {
         $shippingAddress['cart_id'] = $billingAddress['cart_id'] = null;
 
-        if (! empty($data['billing']['isSaveAsAddress'])) {
-            $billingAddress = Arr::except($billingAddress, ['isSaveAsAddress', 'isUseForShipping', 'address_id', 'isNew', 'created_at', 'updated_at']);
+        if (! empty($data['billing']['is_save_as_address'])) {
+            $billingAddress = Arr::except($billingAddress, ['is_save_as_address', 'is_use_for_shipping', 'address_id', 'isNew', 'created_at', 'updated_at']);
             
             $this->customerAddressRepository->updateOrCreate($billingAddress, $billingAddress);
         }
 
-        if (! empty($data['shipping']['isSaveAsAddress'])) {
-            $shippingAddress = Arr::except($shippingAddress, ['isSaveAsAddress', 'isUseForShipping', 'address_id', 'isNew', 'created_at', 'updated_at']);
+        if (! empty($data['shipping']['is_save_as_address'])) {
+            $shippingAddress = Arr::except($shippingAddress, ['is_save_as_address', 'is_use_for_shipping', 'address_id', 'isNew', 'created_at', 'updated_at']);
 
             $this->customerAddressRepository->updateOrCreate($shippingAddress, $shippingAddress);
         }
