+++ packages/Webkul/Shop/src/Resources/views/checkout/onepage/index.blade.php
@@ -9,1255 +9,21 @@
             <div class="container px-[60px] max-lg:px-[30px]">
                 <div class="grid grid-cols-[1fr_auto] gap-[30px]">
                     <div class="grid mt-[30px]">
-                        {{-- Customer addresses component --}}
-                        <v-checkout-addresses ref="vCheckoutAddress"></v-checkout-addresses>
+                        @include('shop::checkout.onepage.address')
 
-                        {{-- Shipping method component --}}
-                        <v-shipping-method ref="vShippingMethod"></v-shipping-method>
+                        @include('shop::checkout.onepage.shipping')
 
-                        {{-- Payment method component --}}
-                        <v-payment-method ref="vPaymentMethod"></v-payment-method>
+                        @include('shop::checkout.onepage.payment')
 
-                        {{-- Order/Review Summary --}}
-                        <v-review-summary ref="vReview"></v-review-summary>
+                        @include('shop::checkout.onepage.review-summary')
                     </div>
                     
-                    {{-- Cart summary --}}
-                    <v-cart-summary
-                        ref="vCartSummary"
-                        :cart="cart"
-                    >
-                    </v-cart-summary>
-                </div>
-            </div>
-        </script>
-    
-        <script type="text/x-template" id="v-checkout-addresses-template">
-            <div>
-                <x-shop::form
-                    v-slot="{ meta, errors, handleSubmit }"
-                    as="div"
-                >
-                    <form @submit="handleSubmit($event, store)">
-                        <div v-if="! address.billing.isNew">
-                            {{-- Addressess list --}}
-                            <x-shop::accordion>
-                                <x-slot:header>
-                                    <div class="flex justify-between items-center">
-                                        <h2 class="text-[26px] font-medium">@lang('Billing Address')</h2>
-                                    </div>
-                                </x-slot:header>
-                            
-                                <x-slot:content>
-                                    <div class="grid mt-[30px] mb-[30px] gap-[20px] grid-cols-2 max-1060:grid-cols-[1fr]">
-                                        <div 
-                                            class="border border-[#e5e5e5] rounded-[12px] p-[0px] max-sm:flex-wrap relative select-none cursor-pointer"
-                                            v-for="(addresses, index) in availableAddresses"
-                                            @change="showBillingMethods(addresses)"
-                                        >
-                                            <v-field
-                                                type="radio"
-                                                name="billing[address_id]"
-                                                :id="'billing_address_id_' + addresses.id"
-                                                :value="addresses.id"
-                                                rules="required"
-                                                v-model="address.billing.address_id"
-                                                class="hidden peer"
-                                            >
-                                            </v-field>
-
-                                            <label 
-                                                class="icon-radio-unselect text-[24px] text-navyBlue absolute right-[20px] top-[20px] peer-checked:icon-radio-select cursor-pointer"
-                                                :for="'billing_address_id_' + addresses.id"
-                                            ></label>
-
-                                            <label 
-                                                :for="'billing_address_id_' + addresses.id"
-                                                class="block p-[20px] rounded-[12px] cursor-pointer"
-                                            >
-                                                <div class="flex justify-between items-center">
-                                                    <p class="text-[16px] font-medium">
-                                                        @{{ addresses.first_name }} @{{ addresses.last_name }}
-                                                        <span v-if="addresses.company_name">(@{{ addresses.company_name }})</span>
-                                                    </p>
-                                                
-                                                    <div 
-                                                        class="flex gap-[25px] items-center"
-                                                        v-if="addresses.default_address"
-                                                    >
-                                                        <div class="m-0 ml-[0px] block mx-auto bg-navyBlue text-white w-max font-medium p-[5px] rounded-[10px] text-center text-[12px]">
-                                                            @lang('Default Address')
-                                                        </div>
-                                                    </div>
-                                                </div>
-
-                                                <p class="text-[#7D7D7D] mt-[25px]">
-                                                    @{{ addresses.address1 }} @{{ addresses.addresses2 }},
-                                                    @{{ addresses.city }}, 
-                                                    @{{ addresses.state }}, @{{ addresses.country }}, 
-                                                    @{{ addresses.postcode }}
-                                                </p>
-                                            </label>
-                                        </div>
-
-                                        <div 
-                                            class="flex justify-center items-center border border-[#e5e5e5] rounded-[12px] p-[20px] max-sm:flex-wrap cursor-pointer"
-                                            @click="addNewBillingAddress"
-                                        >
-                                            <div class="flex gap-x-[10px] items-center">
-                                                <span class="icon-plus text-[30px] p-[30px] border border-black rounded-full"></span>
-                                                <p>@lang('Add new Address')</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </x-slot:content>
-                            </x-shop::accordion>
-                        </div>
-
-                        <div v-else>
-                            <x-shop::accordion>
-                                <x-slot:header>
-                                    <div class="flex justify-between items-center">
-                                        <h2 class="text-[26px] font-medium">@lang('Billing Address')</h2>
-                                    </div>
-                                </x-slot:header>
-                            
-                                <x-slot:content>
-                                    <x-shop::form.control-group>
-
-                                        <div class="flex text-right justify-between">
-                                            <x-shop::form.control-group.label>
-                                                @lang('Company name')
-                                            </x-shop::form.control-group.label>
-                                            
-                                            @if(count(auth('customer')->user()->addresses))
-                                                <a 
-                                                    class="flex"
-                                                    href="javascript:void(0)" 
-                                                    @click="address.billing.isNew = ! address.billing.isNew"
-                                                    v-if="address.billing.isNew"
-                                                >
-                                                    <span class="icon-arrow-left text-[24px]"></span>
-                                                    @lang('Back')
-                                                </a>
-                                            @endif
-                                        </div>
-
-                                        <x-shop::form.control-group.control
-                                            type="text"
-                                            name="billing[company_name]"
-                                            label="Company name"
-                                            placeholder="Company name"
-                                            v-model="address.billing.company_name"
-                                        >
-                                        </x-shop::form.control-group.control>
-                
-                                        <x-shop::form.control-group.error
-                                            control-name="billing[company_name]"
-                                        >
-                                        </x-shop::form.control-group.error>
-                                    </x-shop::form.control-group>
-                
-
-                                    <div class="grid grid-cols-2 gap-x-[20px]">
-                                        <x-shop::form.control-group>
-                                            <x-shop::form.control-group.label class="!mt-[0px]">
-                                                @lang('First name')
-                                            </x-shop::form.control-group.label>
-                    
-                                            <x-shop::form.control-group.control
-                                                type="text"
-                                                name="billing[first_name]"
-                                                label="First name"
-                                                rules="required"
-                                                placeholder="First name"
-                                                v-model="address.billing.first_name"
-                                            >
-                                            </x-shop::form.control-group.control>
-                    
-                                            <x-shop::form.control-group.error
-                                                control-name="billing[first_name]"
-                                            >
-                                            </x-shop::form.control-group.error>
-                                        </x-shop::form.control-group>
-
-                                        <x-shop::form.control-group>
-                                            <x-shop::form.control-group.label class="!mt-[0px]">
-                                                @lang('Last name')
-                                            </x-shop::form.control-group.label>
-                    
-                                            <x-shop::form.control-group.control
-                                                type="text"
-                                                name="billing[last_name]"
-                                                label="Last name"
-                                                rules="required"
-                                                placeholder="Last name"
-                                                v-model="address.billing.last_name"
-                                            >
-                                            </x-shop::form.control-group.control>
-                    
-                                            <x-shop::form.control-group.error
-                                                control-name="billing[last_name]"
-                                            >
-                                            </x-shop::form.control-group.error>
-                                        </x-shop::form.control-group>
-                                    </div>
-                
-                                    <x-shop::form.control-group>
-                                        <x-shop::form.control-group.label class="!mt-[0px]">
-                                            @lang('Email')
-                                        </x-shop::form.control-group.label>
-                
-                                        <x-shop::form.control-group.control
-                                            type="email"
-                                            name="billing[email]"
-                                            rules="required|email"
-                                            label="Email"
-                                            placeholder="email@example.com"
-                                            v-model="address.billing.email"
-                                        >
-                                        </x-shop::form.control-group.control>
-                
-                                        <x-shop::form.control-group.error
-                                            control-name="billing[email]"
-                                        >
-                                        </x-shop::form.control-group.error>
-                                    </x-shop::form.control-group>
-                
-                                    <x-shop::form.control-group>
-                                        <x-shop::form.control-group.label class="!mt-[0px]">
-                                            @lang('Street address')
-                                        </x-shop::form.control-group.label>
-                
-                                        <x-shop::form.control-group.control
-                                            type="text"
-                                            name="billing[address1][]"
-                                            class="text-[14px] shadow appearance-none border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                            rules="required"
-                                            label="Street address"
-                                            placeholder="Street address"
-                                            v-model="address.billing.address1[0]"
-                                        >
-                                        </x-shop::form.control-group.control>
-                
-                                        <x-shop::form.control-group.error
-                                            control-name="billing[address1][]"
-                                        >
-                                        </x-shop::form.control-group.error>
-                                    </x-shop::form.control-group>
-                
-
-                                    <div class="grid grid-cols-2 gap-x-[20px]">
-                                        <x-shop::form.control-group
-                                            class="!mb-4"
-                                        >
-                                            <x-shop::form.control-group.label class="!mt-[0px]">
-                                                @lang('Country')
-                                            </x-shop::form.control-group.label>
-                    
-                                            <x-shop::form.control-group.control
-                                                type="select"
-                                                name="billing[country]"
-                                                class="!text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                rules="required"
-                                                label="Country"
-                                                placeholder="Country"
-                                                v-model="address.billing.country"
-                                            >
-                                                @foreach (core()->countries() as $country)
-                                                    <option value="{{ $country->code }}">{{ $country->name }}</option>
-                                                @endforeach
-                                            </x-shop::form.control-group.control>
-                    
-                                            <x-shop::form.control-group.error
-                                                control-name="billing[country]"
-                                            >
-                                            </x-shop::form.control-group.error>
-                                        </x-shop::form.control-group>
-                
-                                        <x-shop::form.control-group>
-                                            <x-shop::form.control-group.label class="!mt-[0px]">
-                                                @lang('State')
-                                            </x-shop::form.control-group.label>
-                    
-                                            <x-shop::form.control-group.control
-                                                type="text"
-                                                name="billing[state]"
-                                                class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                rules="required"
-                                                label="State"
-                                                placeholder="State"
-                                                v-model="address.billing.state"
-                                                v-if="! isHaveStates('billing')"
-                                            >
-                                            </x-shop::form.control-group.control>
-
-                                            <x-shop::form.control-group.control
-                                                type="select"
-                                                name="billing[state]"
-                                                class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                rules="required"
-                                                label="State"
-                                                placeholder="State"
-                                                v-model="address.billing.state"
-                                                v-if="isHaveStates('billing')"
-                                            >
-                                                <option value="">@lang('Select state')</option>
-
-                                                <option 
-                                                    v-for='(state, index) in states[address.billing.country]' 
-                                                    :value="state.code" 
-                                                >
-                                                    @{{ state.default_name }}
-                                                </option>
-                                            </x-shop::form.control-group.control>
-                    
-                                            <x-shop::form.control-group.error
-                                                control-name="billing[state]"
-                                            >
-                                            </x-shop::form.control-group.error>
-                                        </x-shop::form.control-group>
-                                    </div>
-                
-                                    <div class="grid grid-cols-2 gap-x-[20px]">
-                                        <x-shop::form.control-group>
-                                            <x-shop::form.control-group.label class="!mt-[0px]">
-                                                @lang('City')
-                                            </x-shop::form.control-group.label>
-                
-                                            <x-shop::form.control-group.control
-                                                type="text"
-                                                name="billing[city]"
-                                                class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                rules="required"
-                                                label="City"
-                                                placeholder="City"
-                                                v-model="address.billing.city"
-                                            >
-                                            </x-shop::form.control-group.control>
-                
-                                            <x-shop::form.control-group.error
-                                                control-name="billing[city]"
-                                            >
-                                            </x-shop::form.control-group.error>
-                                        </x-shop::form.control-group>
-                    
-                                        <x-shop::form.control-group>
-                                            <x-shop::form.control-group.label class="!mt-[0px]">
-                                                @lang('Zip/Postcode')
-                                            </x-shop::form.control-group.label>
-                    
-                                            <x-shop::form.control-group.control
-                                                type="text"
-                                                name="billing[postcode]"
-                                                class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                rules="required"
-                                                label="Zip/Postcode"
-                                                placeholder="Zip/Postcode"
-                                                v-model="address.billing.postcode"
-                                            >
-                                            </x-shop::form.control-group.control>
-                    
-                                            <x-shop::form.control-group.error
-                                                control-name="billing[postcode]"
-                                            >
-                                            </x-shop::form.control-group.error>
-                                        </x-shop::form.control-group>
-                                    </div>
-
-                                    <x-shop::form.control-group>
-                                        <x-shop::form.control-group.label class="!mt-[0px]">
-                                            @lang('Telephone')
-                                        </x-shop::form.control-group.label>
-                
-                                        
-                                        <x-shop::form.control-group.control
-                                            type="text"
-                                            name="billing[phone]"
-                                            class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                            rules="required|numeric"
-                                            label="Telephone"
-                                            placeholder="Telephone"
-                                            v-model="address.billing.phone"
-                                        >
-                                        </x-shop::form.control-group.control>
-                
-                                        <x-shop::form.control-group.error
-                                            control-name="billing[phone]"
-                                        >
-                                        </x-shop::form.control-group.error>
-                                    </x-shop::form.control-group>
-            
-                                    <div class="mt-[30px] pb-[15px]">
-                                        <div class="grid gap-[10px]">
-                                            @if ($cart->haveStockableItems())
-                                                <div class="select-none flex gap-x-[15px]">
-                                                    <input 
-                                                        type="checkbox" 
-                                                        name="billing[isUseForShipping]"
-                                                        id="billing[isUseForShipping]" 
-                                                        class="hidden peer"
-                                                        v-model="address.billing.isUseForShipping"
-                                                    >
-
-                                                    <label
-                                                        class="icon-uncheck text-[24px] text-navyBlue peer-checked:icon-check peer-checked:bg-navyBlue peer-checked:rounded-[4px] peer-checked:text-white  cursor-pointer"
-                                                        for="billing[isUseForShipping]"
-                                                    ></label>
-
-                                                    <label for="billing[isUseForShipping]" @click="addNewShippingAddress">@lang('Ship to this address')</label>
-                                                </div>
-                                            @endif
-            
-                                            @auth('customer')
-                                                <div 
-                                                    class="select-none flex gap-x-[15px]"
-                                                    v-if="address.billing.isUseForShipping"
-                                                >
-                                                    <input 
-                                                        type="checkbox"
-                                                        name="billing[isSaveAsAddress]"
-                                                        id="billing[isSaveAsAddress]"
-                                                        class="hidden peer"
-                                                        v-model="address.billing.isSaveAsAddress"
-                                                    >
-
-                                                    <label
-                                                        class="icon-uncheck text-[24px] text-navyBlue peer-checked:icon-check peer-checked:bg-navyBlue peer-checked:rounded-[4px] peer-checked:text-white  cursor-pointer"
-                                                        for="billing[isSaveAsAddress]"
-                                                    ></label>
-
-                                                    <label for="billing[isSaveAsAddress]">@lang('Save this address')</label>
-                                                </div>
-                                            @endauth
-                                        </div>
-                                    </div>
-
-                                    <div 
-                                        class="flex justify-end mt-4 mb-4"
-                                        v-if="address.billing.isUseForShipping"
-                                    >
-                                        <button
-                                            type="submit"
-                                            class="block bg-navyBlue text-white text-base w-max font-medium py-[11px] px-[43px] rounded-[18px] text-center cursor-pointer"
-                                        >
-                                            @lang('Confirm')
-                                        </button>
-                                    </div>
-                                </x-slot:content>
-                                
-                            </x-shop::accordion>
-
-                            @if ($cart->haveStockableItems())
-                                <div 
-                                    v-if="
-                                        ! address.billing.isUseForShipping 
-                                        && address.shipping.isNew
-                                    "
-                                >
-                                    <x-shop::accordion>
-                                        <x-slot:header>
-                                            <div class="flex justify-between items-center">
-                                                <h2 class="text-[26px] font-medium">@lang('Shipping Address')</h2>
-                                            </div>
-                                        </x-slot:header>
-                                    
-                                        <x-slot:content>
-                                            <x-shop::form.control-group>
-                                                <x-shop::form.control-group.label>
-                                                    @lang('Company name')
-                                                </x-shop::form.control-group.label>
-                                            
-                                                <x-shop::form.control-group.control
-                                                    type="text"
-                                                    name="shipping[company_name]"
-                                                    label="Company name"
-                                                    placeholder="Company name"
-                                                    v-model="address.shipping.company_name"
-                                                >
-                                                </x-shop::form.control-group.control>
-                                            
-                                                <x-shop::form.control-group.error
-                                                    control-name="shipping[company_name]"
-                                                >
-                                                </x-shop::form.control-group.error>
-                                            </x-shop::form.control-group>
-
-                                            <div class="grid grid-cols-2 gap-x-[20px]">
-                                                <x-shop::form.control-group>
-                                                    <x-shop::form.control-group.label class="!mt-[0px]">
-                                                        @lang('First name')
-                                                    </x-shop::form.control-group.label>
-                                                
-                                                    <x-shop::form.control-group.control
-                                                        type="text"
-                                                        name="shipping[first_name]"
-                                                        label="First name"
-                                                        rules="required"
-                                                        placeholder="First name"
-                                                        v-model="address.shipping.first_name"
-                                                    >
-                                                    </x-shop::form.control-group.control>
-                                                
-                                                    <x-shop::form.control-group.error
-                                                        control-name="shipping[first_name]"
-                                                    >
-                                                    </x-shop::form.control-group.error>
-                                                </x-shop::form.control-group>
-                                                
-                                                <x-shop::form.control-group>
-                                                    <x-shop::form.control-group.label class="!mt-[0px]">
-                                                        @lang('Last name')
-                                                    </x-shop::form.control-group.label>
-                                                
-                                                    <x-shop::form.control-group.control
-                                                        type="text"
-                                                        name="shipping[last_name]"
-                                                        label="Last name"
-                                                        rules="required"
-                                                        placeholder="Last name"
-                                                        v-model="address.shipping.last_name"
-                                                    >
-                                                    </x-shop::form.control-group.control>
-                                                
-                                                    <x-shop::form.control-group.error
-                                                        control-name="shipping[last_name]"
-                                                    >
-                                                    </x-shop::form.control-group.error>
-                                                </x-shop::form.control-group>
-                                            </div>
-                                            
-                                            <x-shop::form.control-group>
-                                                <x-shop::form.control-group.label class="!mt-[0px]">
-                                                    @lang('Email')
-                                                </x-shop::form.control-group.label>
-                                            
-                                                <x-shop::form.control-group.control
-                                                    type="email"
-                                                    name="shipping[email]"
-                                                    rules="required|email"
-                                                    label="Email"
-                                                    placeholder="email@example.com"
-                                                    v-model="address.shipping.email"
-                                                >
-                                                </x-shop::form.control-group.control>
-                                            
-                                                <x-shop::form.control-group.error
-                                                    control-name="shipping[email]"
-                                                >
-                                                </x-shop::form.control-group.error>
-                                            </x-shop::form.control-group>
-                                            
-                                            <x-shop::form.control-group>
-                                                <x-shop::form.control-group.label class="!mt-[0px]">
-                                                    @lang('Street address')
-                                                </x-shop::form.control-group.label>
-                                            
-                                                <x-shop::form.control-group.control
-                                                    type="text"
-                                                    name="shipping[address1][]"
-                                                    class="text-[14px] shadow appearance-none border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                    rules="required"
-                                                    label="Street address"
-                                                    placeholder="Street address"
-                                                    v-model="address.shipping.address1[0]"
-                                                >
-                                                </x-shop::form.control-group.control>
-                                            
-                                                <x-shop::form.control-group.error
-                                                    control-name="shipping[address1][]"
-                                                >
-                                                </x-shop::form.control-group.error>
-                                            </x-shop::form.control-group>
-                                            
-                                            <div class="grid grid-cols-2 gap-x-[20px]">
-                                                <x-shop::form.control-group
-                                                    class="!mb-4"
-                                                >
-                                                    <x-shop::form.control-group.label class="!mt-[0px]">
-                                                        @lang('Country')
-                                                    </x-shop::form.control-group.label>
-                                                
-                                                    <x-shop::form.control-group.control
-                                                        type="select"
-                                                        name="shipping[country]"
-                                                        class="!text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                        rules="required"
-                                                        label="Country"
-                                                        placeholder="Country"
-                                                        v-model="address.shipping.country"
-                                                    >
-                                                        <option value="">@lang('Select country')</option>
-                                                        @foreach (core()->countries() as $country)
-                                                            <option value="{{ $country->code }}">{{ $country->name }}</option>
-                                                        @endforeach
-                                                    </x-shop::form.control-group.control>
-                                                
-                                                    <x-shop::form.control-group.error
-                                                        control-name="shipping[country]"
-                                                    >
-                                                    </x-shop::form.control-group.error>
-                                                </x-shop::form.control-group>
-                                                
-                                                <x-shop::form.control-group>
-                                                    <x-shop::form.control-group.label class="!mt-[0px]">
-                                                        @lang('State')
-                                                    </x-shop::form.control-group.label>
-                                                
-                                                    <x-shop::form.control-group.control
-                                                        type="text"
-                                                        name="shipping[state]"
-                                                        class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                        rules="required"
-                                                        label="State"
-                                                        placeholder="State"
-                                                        v-model="address.shipping.state"
-                                                        v-if="! isHaveStates('shipping')"
-                                                    >
-                                                    </x-shop::form.control-group.control>
-
-                                                    <x-shop::form.control-group.control
-                                                        type="select"
-                                                        name="shipping[state]"
-                                                        class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                        rules="required"
-                                                        label="State"
-                                                        placeholder="State"
-                                                        v-model="address.shipping.state"
-                                                        v-if="isHaveStates('shipping')"
-                                                    >
-                                                        <option value="">@lang('Select state')</option>
-
-                                                        <option 
-                                                            v-for='(state, index) in states[address.shipping.country]' 
-                                                            :value="state.code"
-                                                        >
-                                                            @{{ state.default_name }}
-                                                        </option>
-                                                    </x-shop::form.control-group.control>
-                                                
-                                                    <x-shop::form.control-group.error
-                                                        control-name="shipping[state]"
-                                                    >
-                                                    </x-shop::form.control-group.error>
-                                                </x-shop::form.control-group>
-                                            </div>
-
-                                            <div class="grid grid-cols-2 gap-x-[20px]">
-                                                <x-shop::form.control-group>
-                                                    <x-shop::form.control-group.label class="!mt-[0px]">
-                                                        @lang('City')
-                                                    </x-shop::form.control-group.label>
-                                                
-                                                    <x-shop::form.control-group.control
-                                                        type="text"
-                                                        name="shipping[city]"
-                                                        class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                        rules="required"
-                                                        label="City"
-                                                        placeholder="City"
-                                                        v-model="address.shipping.city"
-                                                    >
-                                                    </x-shop::form.control-group.control>
-                                                
-                                                    <x-shop::form.control-group.error
-                                                        control-name="shipping[city]"
-                                                    >
-                                                    </x-shop::form.control-group.error>
-                                                </x-shop::form.control-group>
-                                                
-                                                <x-shop::form.control-group>
-                                                    <x-shop::form.control-group.label class="!mt-[0px]">
-                                                        @lang('Zip/Postcode')
-                                                    </x-shop::form.control-group.label>
-                                                
-                                                    <x-shop::form.control-group.control
-                                                        type="text"
-                                                        name="shipping[postcode]"
-                                                        class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                        rules="required"
-                                                        label="Zip/Postcode"
-                                                        placeholder="Zip/Postcode"
-                                                        v-model="address.shipping.postcode"
-                                                    >
-                                                    </x-shop::form.control-group.control>
-                                                
-                                                    <x-shop::form.control-group.error
-                                                        control-name="shipping[postcode]"
-                                                    >
-                                                    </x-shop::form.control-group.error>
-                                                </x-shop::form.control-group>
-                                            </div>
-
-                                            <x-shop::form.control-group>
-                                                <x-shop::form.control-group.label class="!mt-[0px]">
-                                                    @lang('Telephone')
-                                                </x-shop::form.control-group.label>
-                                            
-                                                
-                                                <x-shop::form.control-group.control
-                                                    type="text"
-                                                    name="shipping[phone]"
-                                                    class="text-[14px] bg-white border rounded w-full py-2 px-3 focus:outline-none focus:shadow-outline"
-                                                    rules="required|numeric"
-                                                    label="Telephone"
-                                                    placeholder="Telephone"
-                                                    v-model="address.shipping.phone"
-                                                >
-                                                </x-shop::form.control-group.control>
-                                            
-                                                <x-shop::form.control-group.error
-                                                    control-name="shipping[phone]"
-                                                >
-                                                </x-shop::form.control-group.error>
-                                            </x-shop::form.control-group>
-
-                                            <div class="mt-[30px] border-b-[1px] border-[#E9E9E9] pb-[15px]">
-                                                <div class="grid gap-[10px]">
-                                                    @auth('customer')
-                                                        <div class="select-none flex gap-x-[15px]">
-                                                            <input 
-                                                                type="checkbox"
-                                                                name="billing[isSaveAsAddress]"
-                                                                id="billing[isSaveAsAddress]"
-                                                                class="hidden peer"
-                                                                v-model="address.billing.isSaveAsAddress"
-                                                            >
-                    
-                                                            <span class="icon-uncheck text-[24px] text-navyBlue peer-checked:icon-check peer-checked:bg-navyBlue peer-checked:rounded-[4px] peer-checked:text-white"></span>
-                                                            
-                                                            <label for="billing[isSaveAsAddress]">@lang('Save this address')</label>
-                                                        </div>
-                                                    @endauth
-                                                </div>
-                                            </div>
-
-                                            <div
-                                                class="flex justify-end mt-4 mb-4"
-                                                v-if="! address.billing.isUseForShipping"
-                                            >
-                                                <button
-                                                    type="submit"
-                                                    class="block bg-navyBlue text-white text-base w-max font-medium py-[11px] px-[43px] rounded-[18px] text-center cursor-pointer"
-                                                >
-                                                    @lang('Confirm')
-                                                </button>
-                                            </div>
-                                        </x-slot:content>
-                                    </x-shop::accordion>
-                                </div>
-                            @endif
-                        </div>
-                    </form>
-                </x-shop::form>
-            </div>
-        </script>
-
-        <script type="text/x-template" id="v-shipping-method-template">
-            <div v-if="isShowShippingMethod">
-                <x-shop::accordion>
-                    <x-slot:header>
-                        <div class="flex justify-between mt-2 items-center">
-                            <h2 class="text-[26px] font-medium">@lang('Shipping methods')</h2>
-                        </div>
-                    </x-slot:header>
-
-                    <x-slot:content>
-                        <div class="flex">
-                            <div
-                                class="relative max-w-[218px] select-none m-2"
-                                v-for="shippingMethod in shippingMethods"
-                            >
-                                <div v-for="rate in shippingMethod.rates">
-                                    <input 
-                                        type="radio"
-                                        name="shipping_method"
-                                        :id="rate.method"
-                                        :value="rate.method"
-                                        class="hidden peer"
-                                        @change="save(rate.method)"
-                                    >
-
-                                    <label 
-                                        class="icon-radio-unselect text-[24px] text-navyBlue absolute right-[20px] top-[20px] peer-checked:icon-radio-select cursor-pointer"
-                                        :for="rate.method"
-                                    >
-                                    </label>
-
-                                    <label 
-                                        class="block border border-[#E9E9E9] p-[20px] rounded-[12px] h-[190px] cursor-pointer"
-                                        :for="rate.method"
-                                    >
-                                        <span class="icon-flate-rate text-[60px] text-navyBlue"></span>
-
-                                        <p class="text-[25px] font-semibold mt-[5px]">@{{ rate.base_price }}</p>
-                                        
-                                        <p class="text-[12px] font-medium mt-[10px]">
-                                            <span class="font-medium">@{{ rate.method_title }}</span> - @{{ rate.method_description }}
-                                        </p>
-                                    </label>
-                                </div>
-                            </div>
-                        </div>
-                    </x-slot:content>
-                </x-shop::accordion>
-            </div>
-        </script>
-
-        <script type="text/x-template" id="v-payment-method-template">
-            <div v-if="isShowPaymentMethod">
-                <x-shop::accordion>
-                    <x-slot:header>
-                        <div class="flex justify-between mt-2 items-center">
-                            <h2 class="text-[26px] font-medium">@lang('Payment methods')</h2>
-                        </div>
-                    </x-slot:header>
-
-                    <x-slot:content>
-                        <div class="flex flex-wrap gap-[29px] mb-5">
-                            <div 
-                                class="relative cursor-pointer"
-                                v-for="(payment, index) in paymentMethods"
-                            >
-                                <input 
-                                    type="radio" 
-                                    name="payment[method]" 
-                                    :value="payment.payment"
-                                    :id="payment.method"
-                                    class="hidden peer"    
-                                    @change="paymentMethodSelected(payment)"
-                                >
-
-                                <label :for="payment.method" class="icon-radio-unselect text-[24px] text-navyBlue absolute right-[20px] top-[20px] peer-checked:icon-radio-select cursor-pointer"></label>
-
-                                <label :for="payment.method" class="block border border-[#E9E9E9] p-[20px] rounded-[12px] w-[190px] cursor-pointer">
-                                    <img class="mx-w-[55px] max-h-[45px]" src="{{ bagisto_asset('images/paypal.png') }}" :alt="payment.method_title" :title="payment.method_title">
-                                    <p class="text-[14px] font-semibold mt-[5px]">@{{ payment.method_title }} </p>
-                                    <p class="text-[12px] font-medium mt-[10px]">@{{ payment.description }}</p>
-                                </label>
-                                {{-- Todo implement the additionalDetails --}}
-                                {{-- \Webkul\Payment\Payment::getAdditionalDetails($payment['method'] --}}
-                            </div>
-                        </div>
-                    </x-slot:content>
-                </x-shop::accordion>
-            </div>
-        </script>
-
-        <script type="text/x-template" id="v-review-summary-template">
-            <div v-if="isShowReviewSummary">
-                <x-shop::accordion>
-                    <x-slot:header>
-                        <div class="flex justify-between mt-2 items-center">
-                            <h2 class="text-[26px] font-medium">@lang('Order Summary')</h2>
-                        </div>
-                    </x-slot:header>
-
-                    <x-slot:content>
-                        <div class="flex justify-between">
-                            <div>
-                                <div>
-                                    <b>@lang('Billing Address')</b>
-                                </div>
-                                <ul type="none">
-                                    <li>@{{ reviewCart.billing_address.company_name }}</li>
-                                    <li>@{{ reviewCart.billing_address.first_name }} @{{ reviewCart.billing_address.last_name }}</li>
-                                    <li>@{{ reviewCart.billing_address.address1 }}</li>
-                                    <li>@{{ reviewCart.billing_address.postcode }} @{{ reviewCart.billing_address.city }}</li>
-                                    <li>@{{ reviewCart.billing_address.state }}</li>
-                                    <li>@{{ reviewCart.billing_address.country }} @{{ reviewCart.billing_address.postcode }}</li>
-                                    <li><span class="font-medium">@lang('Contact'):</span> @{{ reviewCart.billing_address.phone }}</li>
-                                </ul>
-                            </div>
-                            <div v-if="
-                                    reviewCart.haveStockableItems 
-                                    && reviewCart.shipping_address
-                                "
-                            >
-                                <div>
-                                    <b>@lang('Shipping Address')</b>
-                                </div>
-                                <ul type="none">
-                                    <li>@{{ reviewCart.shipping_address.company_name }}</li>
-                                    <li>@{{ reviewCart.shipping_address.first_name }} @{{ reviewCart.shipping_address.last_name }}</li>
-                                    <li>@{{ reviewCart.shipping_address.address1 }}</li>
-                                    <li>@{{ reviewCart.shipping_address.postcode }} @{{ reviewCart.shipping_address.city }}</li>
-                                    <li>@{{ reviewCart.shipping_address.state }}</li>
-                                    <li>@{{ reviewCart.shipping_address.country }} @{{ reviewCart.shipping_address.postcode }}</li>
-                                    <li><span class="font-medium">@lang('Contact'):</span> @{{ reviewCart.shipping_address.phone }}</li>
-                                </ul>
-                            </div>
-                        </div>
-                        
-                        <div v-for="item in reviewCart.items" :key="item">
-                            <div class="grid border-b-[1px] border-[#E9E9E9] mt-[40px]">
-                                <div class="flex gap-x-[15px] pb-[20px]">
-                                    <img
-                                        class="max-w-[90px] max-h-[90px] w-[90px] h-[90px] rounded-md"
-                                        :src="item.image.medium_image_url"
-                                        :title="item.name"
-                                        :alt="item.name"
-                                    />
-                                    <div>
-                                        {{-- Need to discussed with (@devansh-sir) about these events --}}
-                                        {{-- {!! view_render_event('bagisto.shop.checkout.name.before', ['item' => $item]) !!} --}}
-
-                                        <p class="text-[26px] font-medium text-navyBlue">@{{ item.name }}</p>
-
-                                        {{-- {!! view_render_event('bagisto.shop.checkout.name.after', ['item' => $item]) !!}
-                                        {!! view_render_event('bagisto.shop.checkout.price.before', ['item' => $item]) !!} --}}
-
-                                        <p class="text-[18px] font-medium mt-[10px]">@{{ item.formatted_total }}</p>
-
-                                        {{-- {!! view_render_event('bagisto.shop.checkout.price.after', ['item' => $item]) !!}
-                                        {!! view_render_event('bagisto.shop.checkout.quantity.before', ['item' => $item]) !!} --}}
-
-                                        <p class="text-[18px]">@{{ item.formatted_price }} X @{{ item.quantity }} (@lang('Quantity'))</p>
-
-                                        {{-- {!! view_render_event('bagisto.shop.checkout.quantity.after', ['item' => $item]) !!}
-
-                                        {!! view_render_event('bagisto.shop.checkout.options.before', ['item' => $item]) !!} --}}
-                                    </div>
-                                </div>
-                                {{-- Need to show additionl data of product --}}
-                            </div>
-                        </div>
-
-                        <div class="flex justify-between">
-                            <div>
-                                <div class="grid gap-[15px] mt-[25px] mb-[30px]">
-                                    <div v-if="reviewCart.haveStockableItems">
-                                        <p class="text-[16px] font-medium">@{{ reviewCart.selected_shipping_rate }}</p>
-                                        <p class="text-[16px]">@{{ reviewCart.selected_shipping_rate_method }}</p>
-                                    </div>
-                                    
-                                    <div>
-                                        <p class="text-[16px]">@lang('Payment method')</p>
-                                        <p class="text-[16px] font-medium">@{{ reviewCart.payment_method }}</p>
-                                    </div>
-
-                                    <div
-                                        class="block bg-navyBlue text-white text-base w-max font-medium py-[11px] px-[43px] rounded-[18px] text-center cursor-pointer"
-                                        @click="placeOrder"
-                                    >
-                                        @lang('Place Order')
-                                    </div>
-                                </div>
-                            </div>
-                            <div>
-                                <div class="grid gap-[15px] mt-[25px] mb-[30px]">
-                                    <div class="flex text-right justify-between">
-                                        <p class="text-[16px]">@lang('Subtotal')</p>
-                                        <p class="text-[16px] font-medium">@{{ reviewCart.base_sub_total  }}</p>
-                                    </div>
-
-                                    <div 
-                                        class="flex text-right justify-between"
-                                        v-for="(amount, index) in reviewCart.base_tax_amounts"
-                                        v-if="parseFloat(reviewCart.base_tax_total)"
-                                    >
-                                        <p class="text-[16px]">Tax (@{{ index }})%</p>
-                                        <p class="text-[16px] font-medium">@{{ amount }}</p>
-                                    </div>
-
-                                    <div 
-                                        class="flex text-right justify-between"
-                                        v-if="reviewCart.selected_shipping_rate"
-                                    >
-                                        <p class="text-[16px] mr-2">@lang('Delivery Charges')</p>
-                                        <p class="text-[16px] font-medium">@{{ reviewCart.selected_shipping_rate }}</p>
-                                    </div>
-
-                                    <div 
-                                        class="flex text-right justify-between"
-                                        v-if="reviewCart.base_discount_amount && parseFloat(reviewCart.base_discount_amount) > 0"
-                                    >
-                                        <p class="text-[16px] mr-2">@lang('Discount amount')</p>
-                                        <p class="text-[16px] font-medium">@{{ reviewCart.formatted_base_discount_amount }}</p>
-                                    </div>
-
-                                    <div class="flex text-right justify-between">
-                                        <p class="text-[16px] mr-2">@lang('Grand total')</p>
-                                        <p class="text-[16px] font-medium"> @{{ reviewCart.base_grand_total }}</p>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </x-slot:content>
-                </x-shop::accordion>
-            </div>
-        </script>
-
-        <script type="text/x-template" id="v-cart-summary-template">
-            <div class="w-[442px] max-w-full pl-[30px] h-max sticky top-[30px] mt-[30px]">
-                <h2 class="text-[26px] font-medium">@lang('Cart Summary')</h2>
-                <div class="grid border-b-[1px] border-[#E9E9E9] mt-[40px]">
-                    <div 
-                        class="flex gap-x-[15px] pb-[20px]"
-                        v-for="item in cart.items"
-                    >
-                        <img
-                            class="max-w-[90px] max-h-[90px] w-[90px] h-[90px] rounded-md"
-                            :src="item.image.medium_image_url"
-                            :title="item.name"
-                            :alt="item.name"
-                        />
-                        <div class="">
-                            <p class="text-[16px] text-navyBlue">@{{ item.name }}</p>
-                            <p class="text-[18px] font-medium mt-[10px]">@{{ item.formatted_total }}</p>
-                            <p class="text-[15px]">@{{ item.formatted_price }} X @{{ item.quantity }} (@lang('Quantity'))</p>
-                        </div>
-                    </div>
-                </div>
-                <div class="grid gap-[15px] mt-[25px] mb-[30px]">
-                    <div class="flex text-right justify-between">
-                        <p class="text-[16px]">@lang('Subtotal')</p>
-                        <p class="text-[16px] font-medium">@{{ cart.base_sub_total  }}</p>
-                    </div>
-
-                    <div 
-                        class="flex text-right justify-between"
-                        v-for="(amount, index) in cart.base_tax_amounts"
-                        v-if="parseFloat(cart.base_tax_total)"
-                    >
-                        <p class="text-[16px]">Tax (@{{ index }})%</p>
-                        <p class="text-[16px] font-medium">@{{ amount }}</p>
-                    </div>
-
-                    <div 
-                        class="flex text-right justify-between"
-                        v-if="cart.selected_shipping_rate"
-                    >
-                        <p class="text-[16px]">@lang('Delivery Charges')</p>
-                        <p class="text-[16px] font-medium">@{{ cart.selected_shipping_rate }}</p>
-                    </div>
-
-                    <div 
-                        class="flex text-right justify-between"
-                        v-if="cart.base_discount_amount && parseFloat(cart.base_discount_amount) > 0"
-                    >
-                        <p class="text-[16px]">@lang('Discount amount')</p>
-                        <p class="text-[16px] font-medium">@{{ cart.formatted_base_discount_amount }}</p>
-                    </div>
-
-                    <v-coupon :is-coupon-applied="cart.coupon_code"></v-coupon>
-
-                    <div class="flex text-right justify-between">
-                        <p class="text-[16px] mr-2">@lang('Grand total')</p>
-                        <p class="text-[16px] font-medium"> @{{ cart.base_grand_total }}</p>
-                    </div>
-                </div>
-            </div>
-        </script>
-
-        <script type="text/x-template" id="v-coupon-template">
-            <div>   
-                <div class="flex text-right justify-between">
-                    <p class="text-[16px] mr-2">@lang('shop::app.checkout.cart.coupon.discount')</p>
-                    <p class="text-[16px] font-medium cursor-pointer">
-                        <x-shop::modal>
-                            <x-slot:toggle>
-                                <span class="text-navyBlue">
-                                    @lang('shop::app.checkout.cart.coupon.apply')
-                                </span>
-                            </x-slot:toggle>
-
-                            <x-slot:header>
-                                @lang('shop::app.checkout.cart.coupon.apply')
-                            </x-slot:header>
-
-                            <x-slot:content>
-
-                                <x-shop::form
-                                    v-slot="{ meta, errors, handleSubmit }"
-                                    as="div"
-                                >
-                                    <form @submit="handleSubmit($event, applyCoupon)">
-                                        <x-shop::form.control-group>
-                                            <x-shop::form.control-group.label>
-                                                @lang('shop::app.checkout.cart.coupon.code')
-                                            </x-shop::form.control-group.label>
-
-                                            <x-shop::form.control-group.control
-                                                type="text"
-                                                name="code"
-                                                placeholder="Enter your code"
-                                                rules="required"
-                                            >
-                                            </x-shop::form.control-group.control>
-
-                                            <x-shop::form.control-group.error
-                                                control-name="code"
-                                            >
-                                            </x-shop::form.control-group.error>
-                                        </x-shop::form.control-group>
-
-                                        <button
-                                            type="submit"
-                                            class="m-0 block bg-navyBlue text-white text-base w-max font-medium py-[11px] px-[43px] rounded-[18px] text-center cursor-pointer"
-                                        >
-                                            @lang('shop::app.customers.account.save')
-                                        </button>
-                                    </form>
-                                </x-shop::form>
-                            </x-slot:content>
-                        </x-shop::modal>
-
-                        <div 
-                            class="text-[12px] font-small flex justify-between items-center"
-                            v-if="isCouponApplied"
-                        >
-                            <p class="text-[12px] mr-2">@lang('Coupon applied')</p>
-                            
-                            <p 
-                                class="text-[16px] font-medium cursor-pointer text-navyBlue"
-                                title="@lang('Applied coupon')"
-                            >
-                                "@{{ isCouponApplied }}"
-                            </p>
-
-                            <span 
-                                class="icon-cancel text-[30px] cursor-pointer"
-                                title="@lang('Remove coupon')"
-                                @click="destroyCoupon"
-                            >
-                            </span>
-                        </div>
-                    </p>
+                    @include('shop::checkout.onepage.cart-summary')
                 </div>
             </div>
         </script>
 
         <script type="module">
-
-            app.component('v-cart-summary', {
-                template: '#v-cart-summary-template',
-                
-                props: ['cart'],
-            })
-
-            app.component('v-coupon', {
-                template: '#v-coupon-template',
-                
-                props: ['isCouponApplied'],
-
-                methods: {
-                    applyCoupon(params) {
-                        this.$axios.post("{{ route('shop.checkout.cart.coupon.apply') }}", params)
-                            .then((response) => {
-                                alert(response.data.data.message)
-                                this.$parent.$parent.getOrderSummary();
-                            })
-                            .catch((error) => {console.log(error);})
-
-                    },
-
-                    destroyCoupon() {
-                        this.$axios.delete("{{ route('shop.checkout.cart.coupon.remove') }}", {
-                                '_token': "{{ csrf_token() }}"
-                            })
-                            .then((response) => {
-
-                                this.$emit('updateOrderSummary')
-
-                                this.$parent.$parent.getOrderSummary();
-                            })
-                            .catch(error => console.log(error));
-                    }
-                }
-            })
-
-            app.component('v-review-summary', {
-                template: '#v-review-summary-template',
-
-                data() {
-                    return  {
-                        isShowReviewSummary: false,
-
-                        reviewCart: {},
-                    }
-                }, 
-
-                methods: {
-                    placeOrder() {
-                        this.$axios.post("{{ route('shop.checkout.save_order') }}", {
-                                '_token': "{{ csrf_token() }}"
-                            })
-                            .then(response => {
-                                if (response.data.success) {
-                                    if (response.data.redirect_url) {
-                                        window.location.href = response.data.redirect_url;
-                                    } else {
-                                        window.location.href = "{{ route('shop.checkout.success') }}";
-                                    }
-                                }
-                            })
-                            .catch(error => console.log(error))
-                    }
-                }
-            })
-
-            app.component('v-payment-method', {
-                template: '#v-payment-method-template',
-
-                data() {
-                    return {
-                        paymentMethods: [],
-
-                        isShowPaymentMethod: false,
-                    }
-                },
-
-                methods: {
-                    paymentMethodSelected(selectedPaymentMethod) {
-                        this.$axios.post("{{ route('shop.checkout.save_payment') }}", {
-                                'payment': selectedPaymentMethod
-                            })
-                            .then(response => {
-                                this.$parent.$refs.vReview.reviewCart = response.data.cart;
-                                
-                                this.$parent.$refs.vReview.isShowReviewSummary = true;
-                            })
-                            .catch(error => console.log(error));
-                    }
-                }
-            })
-
-            app.component('v-shipping-method', {
-                template: '#v-shipping-method-template',
-
-                data() {
-                    return {
-                        shippingMethods: [],
-
-                        isShowShippingMethod: false
-                    }
-                },
-
-                methods: {
-                    save(selectedShippingMethod) {
-                        this.$parent.$refs.vPaymentMethod.isShowPaymentMethod = false;
-
-                        this.$axios.post("{{ route('shop.checkout.save_shipping') }}", {    
-                                shipping_method: selectedShippingMethod,
-                            })
-                            .then(response => {
-                                /*
-                                 * Calling v-checkout's getOrderSummary method component(parent component) getOrdersSummary for update cart summary data.
-                                 */
-                                this.$parent.getOrderSummary();
-
-                                this.$parent.$refs.vPaymentMethod.paymentMethods = response.data.paymentMethods;
-                                    
-                                this.$parent.$refs.vPaymentMethod.isShowPaymentMethod = true;
-
-                                this.$parent.$refs.vReview.isShowReviewSummary = false;
-                            })
-                            .catch(error => {})
-                    },
-                }
-            })
-
             app.component('v-checkout', {
                 template: '#v-checkout-template',
 
@@ -1281,225 +47,6 @@
                     }
                 }
             });
-
-            app.component('v-checkout-addresses', {
-                template: '#v-checkout-addresses-template',
-
-                data() {
-                    return  {
-                        address: {
-                            billing: {
-                                address1: [''],
-
-                                isUseForShipping: true,
-
-                                isNew: false,
-                            },
-
-                            shipping: {
-                                address1: [''],
-
-                                isNew: false,
-                            },
-                        },
-
-                        availableAddresses: [],
-
-                        countries: [],
-
-                        states: [],
-                    }
-                }, 
-                
-                created() {
-                    this.getCustomerAddress();
-
-                    this.getCountryStates();
-
-                    this.getCountries();
-                },
-
-                methods: {
-                    init() {
-                        if (! this.availableAddresses) {
-                            this.address.shipping.isNew = true;
-
-                            this.address.billing.isNew = true;
-                        } else {
-                            this.address.billing.first_name = this.address.shipping.first_name = "{{ auth('customer')->user()->first_name }}";
-                            this.address.billing.last_name = this.address.shipping.last_name = "{{ auth('customer')->user()->last_name }}";
-                            this.address.billing.email = this.address.shipping.email = "{{ auth('customer')->user()->email }}";
-
-                            if (! this.availableAddresses.length) {
-                                this.address.shipping.isNew = true;
-
-                                this.address.billing.isNew = true;
-                            } else {
-                                for (let country in this.countries) {
-                                    for (let code in this.availableAddresses) {
-                                        if (this.availableAddresses[code].country) {
-                                            if (this.availableAddresses[code].country == this.countries[country].code) {
-                                                this.availableAddresses[code]['country'] = this.countries[country].code;
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    },
-
-                    resetPaymentAndShipping() {
-                        this.$parent.$refs.vShippingMethod.isShowShippingMethod = false;
-
-                        this.$parent.$refs.vPaymentMethod.isShowPaymentMethod = false;
-
-                        this.$parent.$refs.vReview.isShowReviewSummary = false;
-                    },
-
-                    assignAddress() {
-                        if (this.availableAddresses.length > 0) {
-                            let address = this.availableAddresses.forEach(address => {
-                                if (address.id == this.address.billing.address_id) {
-                                    this.address.billing.address1 = [address.address1];
-
-                                    if (address.email) {
-                                        this.address.billing.email = address.email;
-                                    }
-
-                                    if (address.first_name) {
-                                        this.address.billing.first_name = address.first_name;
-                                    }
-
-                                    if (address.last_name) {
-                                        this.address.billing.last_name = address.last_name;
-                                    }
-
-                                    if (address.country) {
-                                        this.address.billing.country = address.country;
-                                    }
-                                }
-
-                                if (address.id == this.address.shipping.address_id) {
-                                    this.address.shipping.address1 = [address.address1];
-
-                                    if (address.email) {
-                                        this.address.shipping.email = address.email;
-                                    }
-
-                                    if (address.first_name) {
-                                        this.address.shipping.first_name = address.first_name;
-                                    }
-
-                                    if (address.last_name) {
-                                        this.address.shipping.last_name = address.last_name;
-                                    }
-
-                                    if (address.country) {
-                                        this.address.shipping.country = address.country;
-                                    }
-                                }
-                            });
-                        }
-                    },
-
-                    store() {
-                        this.assignAddress();
-                        
-                        if (this.address.billing.isUseForShipping) {
-                            this.address.shipping = this.address.billing;
-                        }
-                        
-                        if (! this.address.billing.isSaveAsAddress) {
-                            /*
-                             * If the user not click on the save this address checkbox then the address1 array will take first index of address array
-                             */
-                            this.address.billing.address1 = this.address.billing.address1[0];
-
-                            this.availableAddresses.push(this.address.billing);
-                        } else {
-                            this.$axios.post('{{ route("shop.checkout.save_address") }}', this.address)
-                                .then(response => {
-                                    this.$parent.$refs.vShippingMethod.shippingMethods = response.data.data.shippingMethods;
-                                    
-                                    this.$parent.$refs.vShippingMethod.isShowShippingMethod = true;
-                                    
-                                    this.getCustomerAddress();
-                                })
-                                .catch(error => {                 
-                                    console.log(error);
-                                })
-                        }
-
-                        this.address.billing.isNew = false
-                    },
-
-                    getCountries() {
-                        this.$axios.get("{{ route('shop.countries') }}")
-                            .then(response => {
-                                this.countries = response.data.data;
-                            })
-                            .catch(function (error) {});
-                    },
-
-                    getCountryStates() {
-                        this.$axios.get("{{ route('shop.countries.states') }}")
-                            .then(response => {
-                                this.states = response.data.data;
-                            })
-                            .catch(function (error) {});
-                    },
-
-                    getCustomerAddress() {
-                        this.$axios.get("{{ route('api.shop.customers.account.addresses.index') }}")
-                            .then(response => {
-                                this.availableAddresses = response.data.data;
-
-                                this.init();
-                            })
-                            .catch(function (error) {});
-                    },
-
-                    addNewShippingAddress: function () {
-                        this.address.shipping.isNew = true;
-
-                        this.address.shipping.address_id = null;
-
-                        this.resetPaymentAndShipping();
-                    },
-
-                    addNewBillingAddress() {
-                        this.address.billing.isNew = true;
-
-                        this.address.billing.address_id = null;
-
-                        this.$parent.$refs.vShippingMethod.isShowShippingMethod = false;
-
-                        this.$parent.$refs.vPaymentMethod.isShowPaymentMethod = false;
-                    },
-
-                    isHaveStates(addressType) {
-                        if (
-                            this.states[this.address[addressType].country]
-                            && this.states[this.address[addressType].country].length
-                        ) {
-                            return true;
-                        }
-
-                        return false;
-                    },
-
-                    showBillingMethods(address) {
-                        let selectedAddress = this.availableAddresses.find(data => data.id == address.id);
-                        
-                        this.address.billing.isSaveAsAddress = true;
-
-                        this.resetPaymentAndShipping();
-
-                        this.store();
-                    }
-                }
-            })
         </script>
     @endPushOnce
-
 </x-shop::layouts>
