+++ packages/Webkul/Shop/src/Resources/views/products/view.blade.php
@@ -304,10 +304,6 @@
 
                         return value;
                     },
-
-                    onSubmit() {
-                        console.log(this);
-                    }
                 },
             });
         </script>
